fix: move embedded-{io, hal-nb} implementations to `mod trait`
